Create moveclass 2, moving knight
This mini project builds off Mini Project 1. This is the knight's tour.

For this project, take the code from your last project.

Separate the knights class from your source file. Put the class in a header file and a class content file.

The program will do the following:

Ask the user for a starting position.
Places the knight in that starting position and calculates all legal moves. Each move will be numbered.
The user will choose a number and the knight will move to that space and recalculate the number of legal moves.
With each move you will also give the user the option to either start a new game or exit the program.
The program ends when the user wishes.